3835: Added unit tests for probation status badge / html
